Implement GET bookings by date and DELETE booking REST API

+++ src/app/admin/rooms/room-detail/room-detail.component.ts
@@ -29,16 +29,18 @@ export class RoomDetailComponent implements OnInit {
   }
 
   deleteRoom(){
-    this.message = 'deleting...';
-    this.dataService.deleteRoom(this.room.id).subscribe(
-      next => {
-        this.dataChangedEvent.emit();
-        this.router.navigate(['admin', 'rooms'])
-      },
-      error => {
-        this.message = 'Sorry - this room cannot be deleted at this time.';
-      }
+    if (confirm('Are you sure you wish to delete this room?')){
+      this.message = 'deleting...';
+      this.dataService.deleteRoom(this.room.id).subscribe(
+        next => {
+          this.dataChangedEvent.emit();
+          this.router.navigate(['admin', 'rooms']);
+        },
+        error => {
+          this.message = 'Sorry - this room cannot be deleted at this time.';
+        }
       );
+    }
   }
 
 }
